backend/requirements.txt
# Core Django
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-ratelimit==4.1.0
django-rq==3.0.1
django-storages==1.14.4
python-dotenv==1.0.1
PyJWT==2.10.0

# WSGI / static
gunicorn==23.0.0
whitenoise==6.8.2

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Redis / RQ
redis==6.2.0
rq==2.4.1

# GCS + Vertex AI (Veo) + Gemini
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.96.0
google-genai==1.19.0
google-auth==2.40.3

# AI providers
anthropic>=0.42.0
elevenlabs==1.10.0

# HTTP / utilities
requests==2.32.3
httpx==0.28.1
pillow==11.0.0

# Stripe (Phase 5)
stripe==12.5.0

# Cryptography (used by simplejwt)
cryptography==44.0.0
